label-studio-setup/scripts/requirements.txt
label-studio-sdk>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
